Java code for the GP and SP
This code includes the first draft of the multithreaded greenhouse pi
and the server pi which can communicate over localhost right now. The
main changes to be made are: better error detection and logging.

As of now the code works well and does not produce errors. BUT it still
has to be tested on a network.

To run it, open eclipse and click run when you have the
GreenhouseManagement tab open, and then click run again when you have
the GreenhouseMain tab open. This should cause messages to be sent from
both sides and some prints should appear.

Dan